+++ CHANGELOG.md
@@ -5,6 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Fixed
+
+- Use `timezone.now()` instead of `datetime.now()` in `PeriodicFutureTask.save()`.
+- Use `condition` instead of `check` attribute in `CheckConstraint` with `django>=5.1`.
+
+### Added
+
+- Support for Python 3.14.
+- Support for Django 5.2.
+
+### Removed
+
+- Support for Python 3.9.
+- Support for Django 5.0.
+
 ## [1.3.2]
 
 ### Changed
@@ -70,7 +87,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Initial setup.
 
-[Unreleased]: https://github.com/anexia/django-future-tasks/compare/1.3.0...HEAD
+[Unreleased]: https://github.com/anexia/django-future-tasks/compare/1.3.2...HEAD
+[1.3.2]: https://github.com/anexia/django-future-tasks/releases/tag/1.3.2
+[1.3.1]: https://github.com/anexia/django-future-tasks/releases/tag/1.3.1
 [1.3.0]: https://github.com/anexia/django-future-tasks/releases/tag/1.3.0
 [1.2.1]: https://github.com/anexia/django-future-tasks/releases/tag/1.2.1
 [1.2.0]: https://github.com/anexia/django-future-tasks/releases/tag/1.2.0